feat: add shamap module
