+++ catl-hasher.cpp
@@ -20,521 +20,9 @@
 #include "core-types.h"
 #include "log-macros.h"
 #include "logger.h"
+#include "shamap.h"
 #include "utils.h"
 
-
-
-
-// --- End of Logger Implementation ---
-
-
-
-// SHAMap node types
-enum SHAMapNodeType : uint8_t {
-    tnINNER = 1,
-    tnTRANSACTION_NM = 2,
-    tnTRANSACTION_MD = 3,
-    tnACCOUNT_STATE = 4,
-    // TODO: tnUPDATE ? It's context sensitive anyway
-    tnREMOVE = 254,
-    tnTERMINAL = 255
-};
-
-//----------------------------------------------------------
-// Custom Exception Classes
-//----------------------------------------------------------
-class SHAMapException : public std::runtime_error {
-public:
-    explicit SHAMapException(const std::string &message)
-        : std::runtime_error(message) {
-    }
-};
-
-class InvalidDepthException : public SHAMapException {
-public:
-    explicit InvalidDepthException(int depth, size_t maxAllowed)
-        : SHAMapException("Invalid depth (" + std::to_string(depth) +
-                          ") for key in selectBranch. Max allowed: " +
-                          std::to_string(maxAllowed)),
-          depth_(depth), maxAllowed_(maxAllowed) {
-    }
-
-    int depth() const { return depth_; }
-    size_t maxAllowed() const { return maxAllowed_; }
-
-private:
-    int depth_;
-    size_t maxAllowed_;
-};
-
-class InvalidBranchException : public SHAMapException {
-public:
-    explicit InvalidBranchException(int branch)
-        : SHAMapException("Invalid branch index: " + std::to_string(branch)),
-          branch_(branch) {
-    }
-
-    int branch() const { return branch_; }
-
-private:
-    int branch_;
-};
-
-class NullNodeException : public SHAMapException {
-public:
-    explicit NullNodeException(const std::string &context)
-        : SHAMapException("Null node encountered: " + context) {
-    }
-};
-
-class NullItemException : public SHAMapException {
-public:
-    explicit NullItemException()
-        : SHAMapException("Found leaf node with null item") {
-    }
-};
-
-class HashCalculationException : public SHAMapException {
-public:
-    explicit HashCalculationException(const std::string &reason)
-        : SHAMapException("Hash calculation error: " + reason) {
-    }
-};
-
-
-class SHAMapTreeNode;
-class SHAMapInnerNode;
-class SHAMapLeafNode;
-class SHAMap;
-
-int selectBranch(const Key &key, int depth) {
-    int byteIdx = depth / 2;
-    if (byteIdx < 0 || byteIdx >= static_cast<int>(Key::size())) {
-        throw InvalidDepthException(depth, Key::size());
-    }
-    int nibbleIdx = depth % 2;
-    uint8_t byte_val = key.data()[byteIdx];
-    return (nibbleIdx == 0) ? (byte_val >> 4) & 0xF : byte_val & 0xF;
-}
-
-// SHAMapTreeNode
-class SHAMapTreeNode {
-protected:
-    Hash256 hash;
-    bool hashValid = false;
-
-public:
-    void invalidateHash() { hashValid = false; }
-
-    virtual ~SHAMapTreeNode() = default;
-
-    virtual bool isLeaf() const = 0;
-
-    virtual bool isInner() const = 0;
-
-    virtual void updateHash() = 0; // NO LOGGING INSIDE IMPLEMENTATIONS
-    const Hash256 &getHash() {
-        if (!hashValid) {
-            updateHash();
-            hashValid = true;
-        }
-        return hash;
-    }
-};
-
-class SHAMapInnerNode : public SHAMapTreeNode {
-private:
-    std::array<std::shared_ptr<SHAMapTreeNode>, 16> children;
-    uint16_t branchMask = 0;
-    uint8_t depth = 0;
-
-public:
-    SHAMapInnerNode(uint8_t nodeDepth = 0) : depth(nodeDepth) {
-    }
-
-    bool isLeaf() const override { return false; }
-    bool isInner() const override { return true; }
-    uint8_t getDepth() const { return depth; }
-    void setDepth(uint8_t newDepth) { depth = newDepth; }
-
-    void updateHash() override {
-        if (branchMask == 0) {
-            hash = Hash256::zero();
-            hashValid = true;
-            return;
-        }
-        EVP_MD_CTX *ctx = EVP_MD_CTX_new();
-        if (!ctx) throw HashCalculationException("Failed to create EVP_MD_CTX");
-        if (EVP_DigestInit_ex(ctx, EVP_sha512(), nullptr) != 1) {
-            EVP_MD_CTX_free(ctx);
-            throw HashCalculationException("Failed to initialize SHA-512 digest");
-        }
-        auto prefix = HashPrefix::innerNode;
-        if (EVP_DigestUpdate(ctx, &prefix, sizeof(HashPrefix::innerNode)) != 1) {
-            EVP_MD_CTX_free(ctx);
-            throw HashCalculationException("Failed to update digest with prefix");
-        }
-        Hash256 zeroHash = Hash256::zero();
-        for (int i = 0; i < 16; i++) {
-            const uint8_t *hashData = zeroHash.data();
-            if (children[i]) {
-                hashData = children[i]->getHash().data(); // Recursive call might trigger update
-            }
-            if (EVP_DigestUpdate(ctx, hashData, Hash256::size()) != 1) {
-                EVP_MD_CTX_free(ctx);
-                throw HashCalculationException(
-                    "Failed to update digest with child data (branch " + std::to_string(i) + ")");
-            }
-        }
-        std::array<uint8_t, 64> fullHash;
-        unsigned int hashLen = 0;
-        if (EVP_DigestFinal_ex(ctx, fullHash.data(), &hashLen) != 1) {
-            EVP_MD_CTX_free(ctx);
-            throw HashCalculationException("Failed to finalize digest");
-        }
-        EVP_MD_CTX_free(ctx);
-        hash = Hash256(reinterpret_cast<const uint8_t *>(fullHash.data()));
-        hashValid = true;
-    }
-
-    bool setChild(int branch, std::shared_ptr<SHAMapTreeNode> const &child) {
-        if (branch < 0 || branch >= 16) { throw InvalidBranchException(branch); }
-        if (child) {
-            children[branch] = child;
-            branchMask |= (1 << branch);
-            if (child->isInner()) {
-                auto innerChild = std::static_pointer_cast<SHAMapInnerNode>(child);
-                innerChild->setDepth(depth + 1);
-            }
-        } else {
-            children[branch] = nullptr;
-            branchMask &= ~(1 << branch);
-        }
-        invalidateHash(); // Mark self as invalid, not children
-        return true;
-    }
-
-    std::shared_ptr<SHAMapTreeNode> getChild(int branch) const {
-        if (branch < 0 || branch >= 16) { throw InvalidBranchException(branch); }
-        return children[branch];
-    }
-
-    bool hasChild(int branch) const {
-        if (branch < 0 || branch >= 16) { throw InvalidBranchException(branch); }
-        return (branchMask & (1 << branch)) != 0;
-    }
-
-    int getBranchCount() const {
-        int count = 0;
-        for (int i = 0; i < 16; i++) { if (hasChild(i)) count++; }
-        return count;
-    }
-
-    uint16_t getBranchMask() const { return branchMask; }
-
-    std::shared_ptr<SHAMapLeafNode> getOnlyChildLeaf() const {
-        std::shared_ptr<SHAMapLeafNode> resultLeaf = nullptr;
-        int leafCount = 0;
-        for (const std::shared_ptr<SHAMapTreeNode> &childNodePtr: children) {
-            if (childNodePtr) {
-                if (childNodePtr->isInner()) { return nullptr; } // Found inner node
-                leafCount++;
-                if (leafCount == 1) {
-                    resultLeaf = std::static_pointer_cast<SHAMapLeafNode>(childNodePtr);
-                } else {
-                    return nullptr; // Found more than one leaf
-                }
-            }
-        }
-        return resultLeaf; // Returns the leaf if exactly one found, else nullptr
-    }
-};
-
-class SHAMapLeafNode : public SHAMapTreeNode {
-private:
-    std::shared_ptr<MmapItem> item;
-    SHAMapNodeType type;
-
-public:
-    SHAMapLeafNode(std::shared_ptr<MmapItem> i, SHAMapNodeType t)
-        : item(std::move(i)), type(t) { if (!item) { throw NullItemException(); } }
-
-    bool isLeaf() const override { return true; }
-    bool isInner() const override { return false; }
-
-    void updateHash() override {
-        std::array<unsigned char, 4> prefix = {0, 0, 0, 0};
-        auto set = [&prefix](auto &from) { std::memcpy(prefix.data(), from.data(), 4); };
-        switch (type) {
-            case tnTRANSACTION_NM:
-            case tnTRANSACTION_MD: set(HashPrefix::txNode);
-                break;
-            case tnACCOUNT_STATE: default: set(HashPrefix::leafNode);
-                break;
-        }
-        EVP_MD_CTX *ctx = EVP_MD_CTX_new();
-        if (ctx == nullptr) { throw HashCalculationException("Failed to create EVP_MD_CTX"); }
-        if (EVP_DigestInit_ex(ctx, EVP_sha512(), nullptr) != 1) {
-            EVP_MD_CTX_free(ctx);
-            throw HashCalculationException("Failed to initialize SHA-512 digest");
-        }
-        if (EVP_DigestUpdate(ctx, &prefix, sizeof(prefix)) != 1) {
-            EVP_MD_CTX_free(ctx);
-            throw HashCalculationException("Failed to update digest with prefix");
-        }
-        if (EVP_DigestUpdate(ctx, item->slice().data(), item->slice().size()) != 1) {
-            EVP_MD_CTX_free(ctx);
-            throw HashCalculationException("Failed to update digest with item data");
-        }
-        if (EVP_DigestUpdate(ctx, item->key().data(), Key::size()) != 1) {
-            EVP_MD_CTX_free(ctx);
-            throw HashCalculationException("Failed to update digest with item key");
-        }
-        std::array<unsigned char, 64> fullHash;
-        unsigned int hashLen = 0;
-        if (EVP_DigestFinal_ex(ctx, fullHash.data(), &hashLen) != 1) {
-            EVP_MD_CTX_free(ctx);
-            throw HashCalculationException("Failed to finalize digest");
-        }
-        EVP_MD_CTX_free(ctx);
-        hash = Hash256(reinterpret_cast<const uint8_t *>(fullHash.data()));
-        hashValid = true;
-    }
-
-    std::shared_ptr<MmapItem> getItem() const { return item; }
-    SHAMapNodeType getType() const { return type; }
-};
-
-// PathFinder
-class PathFinder {
-private:
-    const Key &targetKey;
-    std::shared_ptr<SHAMapInnerNode> searchRoot;
-    std::vector<std::shared_ptr<SHAMapInnerNode> > inners;
-    std::vector<int> branches;
-    std::shared_ptr<SHAMapLeafNode> foundLeaf = nullptr;
-    bool leafKeyMatches = false;
-    int terminalBranch = -1;
-
-    void findPath(std::shared_ptr<SHAMapInnerNode> root) {
-        if (!root) { throw NullNodeException("PathFinder: null root node"); }
-        searchRoot = root;
-        foundLeaf = nullptr;
-        leafKeyMatches = false;
-        terminalBranch = -1;
-        std::shared_ptr<SHAMapInnerNode> currentInner = root;
-        while (true) {
-            int branch = selectBranch(targetKey, currentInner->getDepth());
-            std::shared_ptr<SHAMapTreeNode> child = currentInner->getChild(branch);
-            if (!child) {
-                terminalBranch = branch;
-                inners.push_back(currentInner);
-                break;
-            }
-            if (child->isLeaf()) {
-                terminalBranch = branch;
-                inners.push_back(currentInner);
-                foundLeaf = std::static_pointer_cast<SHAMapLeafNode>(child);
-                if (foundLeaf->getItem()) {
-                    leafKeyMatches = (foundLeaf->getItem()->key() == targetKey);
-                } else { throw NullItemException(); }
-                break;
-            }
-            inners.push_back(currentInner);
-            branches.push_back(branch);
-            currentInner = std::static_pointer_cast<SHAMapInnerNode>(child);
-        }
-    }
-
-public:
-    PathFinder(std::shared_ptr<SHAMapInnerNode> &root, const Key &key) : targetKey(key) { findPath(root); }
-    bool hasLeaf() const { return foundLeaf != nullptr; }
-    bool didLeafKeyMatch() const { return leafKeyMatches; }
-    bool endedAtNullBranch() const { return foundLeaf == nullptr && terminalBranch != -1; }
-    std::shared_ptr<const SHAMapLeafNode> getLeaf() const { return foundLeaf; }
-    std::shared_ptr<SHAMapLeafNode> getLeafMutable() { return foundLeaf; }
-    std::shared_ptr<SHAMapInnerNode> getParentOfTerminal() { return inners.empty() ? nullptr : inners.back(); }
-
-    std::shared_ptr<const SHAMapInnerNode> getParentOfTerminal() const {
-        return inners.empty() ? nullptr : inners.back();
-    }
-
-    int getTerminalBranch() const { return terminalBranch; }
-
-    void dirtyPath() const {
-        for (auto &inner: inners) { inner->invalidateHash(); }
-    }
-
-    void collapsePath() {
-        if (inners.size() <= 1) return;
-        std::shared_ptr<SHAMapLeafNode> onlyChild = nullptr;
-        auto innermost = inners.back();
-        onlyChild = innermost->getOnlyChildLeaf();
-        for (int i = static_cast<int>(inners.size()) - 2; i >= 0; --i) {
-            auto inner = inners[i];
-            int branch = branches[i];
-            if (onlyChild) { inner->setChild(branch, onlyChild); }
-            onlyChild = inner->getOnlyChildLeaf();
-            if (!onlyChild) break;
-        }
-    }
-};
-
-
-// SHAMap class (Integrate Logging for errors/debug)
-class SHAMap {
-private:
-    std::shared_ptr<SHAMapInnerNode> root;
-    SHAMapNodeType nodeType;
-
-public:
-    SHAMap(SHAMapNodeType type = tnACCOUNT_STATE) : nodeType(type) {
-        root = std::make_shared<SHAMapInnerNode>(0); // Root has depth 0
-        LOGD("SHAMap created with type: ", static_cast<int>(type));
-    }
-
-    // Added for completeness, might not be used externally in this tool
-    Hash256 getChildHash(int ix) {
-        try {
-            if (!root) return Hash256::zero();
-            auto child = root->getChild(ix);
-            if (child) {
-                return child->getHash();
-            } else {
-                return Hash256::zero();
-            }
-        } catch (const InvalidBranchException &e) {
-            LOGW("Attempted to get child hash for invalid branch ", ix, " from root: ", e.what());
-            return Hash256::zero();
-        }
-    }
-
-    // TODO: add a custom return type for this, to distinguish between adding and updating
-    // fail/added/updated <- return type, need more than just bool, need a tribool
-    bool addItem(std::shared_ptr<MmapItem> &item, bool allowUpdate = true) {
-        if (!item) {
-            LOGW("Attempted to add null item to SHAMap.");
-            return false;
-        }
-        LOGD_KEY("Attempting to add item with key: ", item->key());
-
-        try {
-            PathFinder pathFinder(root, item->key());
-
-            if (pathFinder.endedAtNullBranch() || (
-                    pathFinder.hasLeaf() && pathFinder.didLeafKeyMatch() && allowUpdate)) {
-                auto parent = pathFinder.getParentOfTerminal();
-                int branch = pathFinder.getTerminalBranch();
-                if (!parent) { throw NullNodeException("addItem: null parent node (should be root)"); }
-
-                LOGD("Adding/Updating leaf at depth ", parent->getDepth() + 1, " branch ", branch);
-                auto newLeaf = std::make_shared<SHAMapLeafNode>(item, nodeType);
-                parent->setChild(branch, newLeaf);
-                pathFinder.dirtyPath();
-                return true;
-            }
-
-            if (pathFinder.hasLeaf() && !pathFinder.didLeafKeyMatch()) {
-                LOGD_KEY("Handling collision for key: ", item->key());
-                auto parent = pathFinder.getParentOfTerminal();
-                int branch = pathFinder.getTerminalBranch();
-                if (!parent) { throw NullNodeException("addItem collision: null parent node (should be root)"); }
-                auto existingLeaf = pathFinder.getLeafMutable();
-                auto existingItem = existingLeaf->getItem();
-                if (!existingItem) { throw NullItemException(); /* Should be caught by leaf constructor */ }
-
-                std::shared_ptr<SHAMapInnerNode> currentParent = parent;
-                int currentBranch = branch;
-                uint8_t currentDepth = parent->getDepth() + 1; // Start depth below parent
-
-                // Create first new inner node to replace the leaf
-                auto newInner = std::make_shared<SHAMapInnerNode>(currentDepth);
-                parent->setChild(currentBranch, newInner);
-                currentParent = newInner;
-
-                while (currentDepth < 64) {
-                    // Max depth check
-                    int existingBranch = selectBranch(existingItem->key(), currentDepth);
-                    int newBranch = selectBranch(item->key(), currentDepth);
-
-                    if (existingBranch != newBranch) {
-                        LOGD("Collision resolved at depth ", currentDepth, ". Placing leaves at branches ",
-                             existingBranch, " and ", newBranch);
-                        auto newLeaf = std::make_shared<SHAMapLeafNode>(item, nodeType);
-                        currentParent->setChild(existingBranch, existingLeaf);
-                        currentParent->setChild(newBranch, newLeaf);
-                        break; // Done
-                    } else {
-                        // Collision continues, create another inner node
-                        LOGD("Collision continues at depth ", currentDepth, ", branch ", existingBranch,
-                             ". Descending further.");
-                        auto nextInner = std::make_shared<SHAMapInnerNode>(currentDepth + 1);
-                        currentParent->setChild(existingBranch, nextInner);
-                        currentParent = nextInner;
-                        currentDepth++;
-                    }
-                }
-                if (currentDepth >= 64) {
-                    throw SHAMapException("Maximum SHAMap depth reached during collision resolution for key: " +
-                                          item->key().toString());
-                }
-
-                pathFinder.dirtyPath();
-                return true;
-            }
-
-            // Should ideally not be reached if PathFinder logic is correct
-            LOGE("Unexpected state in addItem for key: ", item->key().toString(), ". PathFinder logic error?");
-            throw SHAMapException("Unexpected state in addItem - PathFinder logic error");
-        } catch (const SHAMapException &e) {
-            LOGE("Error adding item with key ", item->key().toString(), ": ", e.what());
-            return false;
-        }
-        catch (const std::exception &e) {
-            LOGE("Standard exception adding item with key ", item->key().toString(), ": ", e.what());
-            return false;
-        }
-    }
-
-    bool removeItem(const Key &key) {
-        LOGD_KEY("Attempting to remove item with key: ", key);
-        try {
-            PathFinder pathFinder(root, key);
-
-            if (!pathFinder.hasLeaf() || !pathFinder.didLeafKeyMatch()) {
-                LOGD_KEY("Item not found for removal, key: ", key);
-                return false; // Item not found
-            }
-
-            auto parent = pathFinder.getParentOfTerminal();
-            int branch = pathFinder.getTerminalBranch();
-            if (!parent) { throw NullNodeException("removeItem: null parent node (should be root)"); }
-
-            LOGD("Removing leaf at depth ", parent->getDepth() + 1, " branch ", branch);
-            parent->setChild(branch, nullptr); // Remove the leaf
-            pathFinder.dirtyPath();
-            pathFinder.collapsePath(); // Compress path if possible
-            LOGD_KEY("Item removed successfully, key: ", key);
-            return true;
-        } catch (const SHAMapException &e) {
-            LOGE("Error removing item with key ", key.toString(), ": ", e.what());
-            return false;
-        }
-        catch (const std::exception &e) {
-            LOGE("Standard exception removing item with key ", key.toString(), ": ", e.what());
-            return false;
-        }
-    }
-
-    Hash256 getHash() const {
-        if (!root) {
-            LOGW("Attempting to get hash of a null root SHAMap.");
-            return Hash256::zero();
-        }
-        // getHash() inside the node handles lazy calculation
-        return root->getHash();
-    }
-};
 
 class CATLHasher {
 private:
